+++ changelog.md
@@ -32,6 +32,7 @@
 * `FIX` [#1051](https://github.com/sumneko/lua-language-server/issues/1051)
 * `FIX` [#1072](https://github.com/sumneko/lua-language-server/issues/1072)
 * `FIX` [#1077](https://github.com/sumneko/lua-language-server/issues/1077)
+* `FIX` [#1088](https://github.com/sumneko/lua-language-server/issues/1088)
 * `FIX` runtime errors
 
 ## 3.1.0