+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 3.2.0
+`2022-4-25`
 * `NEW` supports infer of callback parameter
   ```lua
   ---@type string[]